Use u16 for msg_id
Modem accepts 2-byte MSG IDs, we can use that to our advantage

+++ common/src/at/mqtt.rs
@@ -9,12 +9,12 @@ pub enum MqttPubStatus {
 
 #[derive(PartialEq, Eq)]
 pub struct MqttPublishReport {
-    pub msg_id: u8,
+    pub msg_id: u16,
     pub status: MqttPubStatus,
 }
 
 impl MqttPublishReport {
-    pub fn from_bg77_qmtpub(msg_id: u8, status: u8, retries: Option<&u8>) -> Self {
+    pub fn from_bg77_qmtpub(msg_id: u16, status: u8, retries: Option<&u8>) -> Self {
         let status = match status {
             0 => MqttPubStatus::Published,
             1 => MqttPubStatus::Retrying(*retries.unwrap_or(&0)),
@@ -24,7 +24,7 @@ impl MqttPublishReport {
         Self { msg_id, status }
     }
 
-    pub fn mqtt_error(msg_id: u8) -> Self {
+    pub fn mqtt_error(msg_id: u16) -> Self {
         Self {
             msg_id,
             status: MqttPubStatus::MqttError,
